+++ libs/cardano-ledger-api/test/Tests.hs
@@ -13,6 +13,7 @@ import Data.Default (def)
 import Test.Cardano.Ledger.Allegra.Binary.Annotator ()
 import Test.Cardano.Ledger.Alonzo.Binary.Annotator ()
 import qualified Test.Cardano.Ledger.Api.State.Imp.QuerySpec as ImpQuery (spec)
+import qualified Test.Cardano.Ledger.Api.State.Query.GoldenSpec as QueryGolden (spec)
 import qualified Test.Cardano.Ledger.Api.State.QuerySpec as StateQuery (spec)
 import qualified Test.Cardano.Ledger.Api.Tx as Tx (spec)
 import qualified Test.Cardano.Ledger.Api.Tx.Body as TxBody (spec)
@@ -37,6 +38,7 @@ apiSpec =
       TxBody.spec
     describe "State" $ do
       StateQuery.spec
+      QueryGolden.spec
     describe "Imp" $
       ImpQuery.spec @ConwayEra
     describe "Upgrade" $ do